+++ .config/pre-commit-config.yaml
@@ -17,7 +17,7 @@ repos:
       - id: ruff-format
         args: [--config, .config/ruff.toml]
   - repo: https://github.com/astral-sh/uv-pre-commit
-    rev: 0.5.29
+    rev: 0.6.1
     hooks:
       # Run the pip compile
       - id: pip-compile
