+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.10.40] = 2025-02-20
+
+### Changed
+
+- Humanitarian needs reads from HAPI dataset
+
 ## [0.10.39] = 2025-02-18
 
 ### Changed
