+++ pyproject.toml
@@ -38,8 +38,8 @@ dependencies = [
     "hdx-python-api>= 6.3.8",
     "hdx-python-country>= 3.8.8",
     "hdx-python-database[postgresql]>= 1.3.5",
-    "hdx-python-scraper>= 2.6.2",
-    "hdx-python-utilities>= 3.8.3",
+    "hdx-python-scraper>= 2.6.3",
+    "hdx-python-utilities>= 3.8.4",
     "libhxl",
     "sqlalchemy"
 ]
