+++ requirements.txt
@@ -67,15 +67,15 @@ hdx-python-country==3.8.8
     #   hdx-python-scraper
 hdx-python-database==1.3.5
     # via hapi-pipelines (pyproject.toml)
-hdx-python-scraper==2.6.2
+hdx-python-scraper==2.6.3
     # via hapi-pipelines (pyproject.toml)
-hdx-python-utilities==3.8.3
+hdx-python-utilities==3.8.4
     # via
     #   hapi-pipelines (pyproject.toml)
     #   hdx-python-api
     #   hdx-python-country
     #   hdx-python-scraper
-humanize==4.12.0
+humanize==4.12.1
     # via frictionless
 identify==2.6.7
     # via pre-commit
